Several improvements to collision mode
When in collision mode, now only the collision mesh of the selected room is forced to be visible (because forcing every collision mesh in the dungeon visible at the same time makes it impossible to see anything).

Also a minor bug fix where, if you're in collision mode and select a room, then switch to actor mode and select the stage, then switch back to collision mode, collision mode would act like you still had the room selected.