Fix crash when clicking in actor mode if root dir not set
The cause of this was that m_objRender was never initialized for objects with a custom PostLoad function. m_objRender would normally be initialized in VisibleDOMNode's PostLoad, but because most child classes don't call base.PostLoad(), it was never actually being initialized for those (except ones with a custom m_objRender).

fixes #17