Significantly changed how trigger regions are displayed
Instead of the huge region always being invisible, instead only a small cube in the center is visible now. When you select the small cube, the entire region becomes visible. This is to prevent tons of overlapping regions from always being visible and overlapping, which makes everything very difficult to see.