Implement new feature to auto-populate subtype specific default params
e.g. If you select a golden crate from the actor creator window, it will now set the type to golden crate from the start so the user doesn't get a wooden crate and have to change a param manually.
Only a few actors have this implemented so far though, need to do more.