Add hack to fix collision disappearing sometimes
On 64 bit Winditor builds, parts of the floor of the Ura/Ura_jimen_B mesh in Dragon Roost Island's DZB file would disappear when saving with no changes.
The exact cause is unknown but is probably related to different floating point rounding errors between 32 and 64 bit builds.
This hack just expands the calculated collision group bounds by a little bit in all 6 directions in the hopes of avoiding this bug in the future.